improve diff coverage
